fix: kyle's untested total_bytes_consumed code

+++ src/netrics/netson.py
@@ -517,11 +517,11 @@ class Measurements:
                 if ul_dl == "ul":
                     sum_sent = res["SumSent"]
                     self.results[key]['avg_sum_sent'] = float(sum_sent) / 12
-                    self.results['total_bytes_consumed'] += float(sum_sent)
+                    self.results['total_bytes_consumed'] += int(sum_sent)
                 else:
                     sum_recv = res["SumRecv"]
                     self.results[key]['avg_sum_recv'] = float(sum_recv) / 12
-                    self.results['total_bytes_consumed'] += float(sum_recv)
+                    self.results['total_bytes_consumed'] += int(sum_recv)
 
                 tcp_rates = []
                 icmp_rates = []
